+++ examples/resources/cockroach_allow_list/import.sh
@@ -1,0 +1,2 @@
+# format: <cluster id>:<cidr ip>/<cidr mask>
+terraform import cockroach_allow_list.home_office 1f69fdd2-600a-4cfc-a9ba-16995df0d77d:123.123.1.1/32