+++ examples/resources/cockroach_private_endpoint_connection/import.sh
@@ -1,0 +1,2 @@
+# format: <cluster id>:<connection id>
+terraform import cockroach_private_endpoint_connection.resource_name 1f69fdd2-600a-4cfc-a9ba-16995df0d77d:vpce-0c1308d7312217abc