+++ examples/resources/cockroach_private_endpoint_trusted_owner/import.sh
@@ -1,0 +1,2 @@
+# format: <cluster id>:<owner id>
+terraform import cockroach_private_endpoint_trusted_owner.resource_name 1f69fdd2-600a-4cfc-a9ba-16995df0d77d:e50aa10d-1a16-4be8-85e6-4c18221daa49